Avoid specifying input pins; we can just refer to I/O ports on spy buffers
I had thought we might need to explicitly toggle the wires connected
to the spy buffers in the toplevel wrapper, but it turns out that's
not necessary.

Instead of needing to explicitly specify the read enable
when creating a spy buffer monitor, for instance, this commit makes
it possible to just specify the path _to_ the spy buffer. The monitor
then knows that the read enable pin is called "read_enable" on every
spy buffer instance.