Add test to read binary input file; add creation of binary log files
This commit adds the creation of binary event files to the output monitor;
these are produced using the same 9-byte (1 metadata byte + a long) format
used in the system-simulation repository.

This commit also adds a test which loads one of these files as _input_
and reads events from it into the spy buffer(s).